scertecd: retry DNS creation probe errors more quickly
The default probe interval for DNS record creation is 1 hour.  If one fails,
even if it was some transient AWS thing, the possible successq will be an hour
later, and the last success will look stale.

This makes the probe alerts occasionally (slightly) noisy.  To mitigate that,
retry failures sooner.